PT-1951 Patient measurements - UI improvements move static methods into separate utility class
